0325 release version 4

+++ app/src/main/java/com/umcapplunching/cocktail_dakk/ui/main/mainrecommand/MainrecommandFragment.kt
@@ -2,10 +2,14 @@ package com.umcapplunching.cocktail_dakk.ui.main.mainrecommand
 
 import android.content.Context
 import android.content.SharedPreferences
+import android.os.Bundle
+import android.view.LayoutInflater
 import android.view.View
+import android.view.ViewGroup
 import android.view.WindowManager
 import androidx.activity.OnBackPressedCallback
 import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
 import androidx.viewpager2.widget.ViewPager2
 import com.umcapplunching.cocktail_dakk.R
 import com.umcapplunching.cocktail_dakk.data.entities.cocktaildata_db.CocktailDatabase
@@ -19,18 +23,31 @@ import com.umcapplunching.cocktail_dakk.ui.main.mainrecommand.MainrecService.Mai
 import com.umcapplunching.cocktail_dakk.utils.getaccesstoken
 import kotlinx.coroutines.launch
 import kotlin.math.abs
+import android.app.Activity
+import android.util.Log
 
-class MainrecommandFragment : BaseFragment<FragmentMainrecommandBinding>(FragmentMainrecommandBinding::inflate), MainrecView {
+
+//class MainrecommandFragment : BaseFragment<FragmentMainrecommandBinding>(FragmentMainrecommandBinding::inflate), MainrecView {
+class MainrecommandFragment : Fragment() ,MainrecView {
 
     private lateinit var CocktailDB : CocktailDatabase
     val mainrecService = MainrecService()
     lateinit var bannerAdapter : BannerViewpagerAdapter
     private lateinit var callback: OnBackPressedCallback
+    private var _binding : FragmentMainrecommandBinding? = null
+    private val binding get() = _binding!!
 
-    override fun initAfterBinding() {
-        requireActivity().window.setFlags(
-                    WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE,
-                    WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE)
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentMainrecommandBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onStart() {
+        super.onStart()
 
         //메인화면 서버연결
         mainrecService.setmainrecView(this)
@@ -45,9 +62,9 @@ class MainrecommandFragment : BaseFragment<FragmentMainrecommandBinding>(Fragmen
         }
         binding.mainRecVp.offscreenPageLimit = 1 // 몇 개의 페이지를 미리 로드 해둘것인지
 
-        launch {
+//        launch {
             mainrecService.mainRec(getaccesstoken(requireContext()))
-        }
+//        }
         val spf = activity?.getSharedPreferences("currenttab", AppCompatActivity.MODE_PRIVATE)
         val editor: SharedPreferences.Editor = spf?.edit()!!
         editor.putInt("currenttab", 1)
@@ -59,42 +76,35 @@ class MainrecommandFragment : BaseFragment<FragmentMainrecommandBinding>(Fragmen
 
     override fun onMainrecSuccess(mainrecList : Mainrec) {
         //DB설정
-        CocktailDB.MainrecDao().deleteAllCocktail()
-        requireActivity().window.clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE)
+//        requireActivity().window.clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE)
+        val activity: Activity? = activity
+        if ( isAdded() && activity != null) {
+            CocktailDB.MainrecDao().deleteAllCocktail()
+            binding.mainRecLoadingPb.visibility = View.GONE
+            binding.mainRecTv.text = mainrecList.nickname + resources.getString(R.string.main_coktailrecommand)
 
-        //해결해야함
-//        requireActivity().runOnUiThread(object : Runnable{
-//            override fun run() {
-//            }
-//        })
-
-//        android.os.Handler().postDelayed(object : Runnable{
-//            override fun run() {
-//            }
-//        },500)
-
-        binding.mainRecLoadingPb.visibility = View.GONE
-        binding.mainRecTv.text = mainrecList.nickname + resources.getString(R.string.main_coktailrecommand)
-
-        for (i in mainrecList.userRecommendationLists.indices){
-            bannerAdapter.addFragment(mainrecList.userRecommendationLists[i].cocktailInfoId,mainrecList.userRecommendationLists[i].cocktailImageURL)
-            //DB에 저장
-            CocktailDB.MainrecDao().insert(
-            Cocktail_Mainrec(
-                mainrecList.userRecommendationLists[i].koreanName,mainrecList.userRecommendationLists[i].cocktailInfoId
-            ))
+            for (i in mainrecList.userRecommendationLists.indices){
+                bannerAdapter.addFragment(mainrecList.userRecommendationLists[i].cocktailInfoId,mainrecList.userRecommendationLists[i].cocktailImageURL)
+                //DB에 저장
+                CocktailDB.MainrecDao().insert(
+                    Cocktail_Mainrec(
+                        mainrecList.userRecommendationLists[i].koreanName,mainrecList.userRecommendationLists[i].cocktailInfoId
+                    ))
+            }
+            binding.mainRecVp.adapter = bannerAdapter
+            binding.mainRecVp.orientation = ViewPager2.ORIENTATION_HORIZONTAL
+            binding.mainRecIndicator.setViewPager2(binding.mainRecVp)
         }
-        binding.mainRecVp.adapter = bannerAdapter
-        binding.mainRecVp.orientation = ViewPager2.ORIENTATION_HORIZONTAL
-        binding.mainRecIndicator.setViewPager2(binding.mainRecVp)
 
     }
 
     override fun onSignUpFailure(code: Int, message: String) {
-        binding.mainRecLoadingPb.visibility = View.GONE
-        requireActivity().window.clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE)
-        if (code==5000){
-                (activity as MainActivity).TokenrefreshInMain()
+        val activity: Activity? = activity
+            if ( isAdded() && activity != null) {
+            binding.mainRecLoadingPb.visibility = View.GONE
+            if (code==5000){
+                    (activity as MainActivity).TokenrefreshInMain()
+            }
         }
     }
 
@@ -104,7 +114,7 @@ class MainrecommandFragment : BaseFragment<FragmentMainrecommandBinding>(Fragmen
             override fun handleOnBackPressed() {
             }
         }
-        requireActivity().onBackPressedDispatcher.addCallback(this, callback)
+        //requireActivity().onBackPressedDispatcher.addCallback(this, callback)
     }
 
     override fun onDetach() {
